requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2